[level 1] Title: x만큼 간격이 있는 n개의 숫자, Time: 0.06 ms, Memory: 75.4 MB -BaekjoonHub

+++ 프로그래머스/1/12954. x만큼 간격이 있는 n개의 숫자/x만큼 간격이 있는 n개의 숫자.java
@@ -1,17 +1,10 @@
 class Solution {
-    public long[] solution(long x, int n) {
-        // long[] answer = {};
+    public long[] solution(int x, int n) {
         long[] answer = new long[n];
         
         for(int i=0; i<n; i++) {
-            answer[i] = x * (i+1);
+            answer[i] = (long)x * (i+1); // 오버플로우가 일어날 수 있기 때문에 long 형으로 저장해주어야 함
         }
-        
-        // int i = 0;
-        // while(i < n) {
-        //     answer[i] = x * (i+1);
-        //     ++i;
-        // }
         
         return answer;
     }
